(+) Исправление доменных моделей

+++ src/backend/dotnet/Domain/Models/Feedback.cs
@@ -4,7 +4,10 @@ public class Feedback
 {
     public Feedback() { }
     
-    public Feedback(Guid id, Guid registationId, string comment, double rate)
+    public Feedback(Guid id, 
+        Guid registationId, 
+        string comment, 
+        int rate)
     {
         Id = id;
         Comment = comment;
@@ -15,6 +18,6 @@ public class Feedback
     public Guid Id { get; set; }
     public Guid RegistationId { get; set; }
     public string Comment { get; set; }
-    public double Rate { get; set; }
+    public int Rate { get; set; }
 }
