+++ src/backend/dotnet/Domain/Models/Location.cs
@@ -4,7 +4,11 @@ public class Location
 {
     public Location() { }
     
-    public Location(Guid id, string title, string description, double cost, int capacity)
+    public Location(Guid id, 
+        string title, 
+        string description, 
+        decimal cost, 
+        int capacity)
     {
         Id = id;
         Title = title;
@@ -16,6 +20,6 @@ public class Location
     public Guid Id { get; set; }
     public string Title { get; set; }
     public string Description { get; set; }
-    public double Cost { get; set; }
+    public decimal Cost { get; set; }
     public int Capacity { get; set; }
 }
