+++ src/backend/dotnet/Domain/Models/MenuItems.cs
@@ -4,7 +4,9 @@ public class MenuItems
 {
     public MenuItems() { }
     
-    public MenuItems(Guid menuId, Guid itemId, double amount)
+    public MenuItems(Guid menuId, 
+        Guid itemId, 
+        int amount)
     {
         MenuId = menuId;
         ItemId = itemId;
@@ -13,5 +15,5 @@ public class MenuItems
     
     public Guid MenuId { get; set; }
     public Guid ItemId { get; set; }
-    public double Amount { get; set; }
+    public int Amount { get; set; }
 }
